app/requirements.txt
asgiref==3.7.1
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
Markdown==3.4.3
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
